+++ packages/google_mobile_ads/CHANGELOG.md
@@ -17,6 +17,14 @@
   * See the app open [example app](https://github.com/googleads/googleads-mobile-flutter/tree/master/packages/app_open_example) for a reference
     on how to use the new API.
     
+* Adds a new parameter `extras` to `AdRequest` and `AdManagerAdRequest`.
+  * This can be used to pass additional signals to the AdMob adapter, such as
+    [CCPA](https://developers.google.com/admob/android/ccpa) signals.
+  * For example, to notify Google that [RDP](https://developers.google.com/admob/android/ccpa#rdp_signal)
+    should be enabled when constructing an ad request:
+    ```dart
+      AdRequest request = AdRequest(extras: {'rdp': '1'});
+    ```
 ## 0.13.6
 
 * Partial fix for [#265](https://github.com/googleads/googleads-mobile-flutter/issues/265).